[fix] add `null-coalescing operator` to `setViews` method
API (`instagram plugin`) returns `reels` with `view count` equals to `null`. But in `setViews` function in `reelsBaseHydrator` `view count` can't be a `null` (only `int` can be set). This leads to `exception`.

This update just adds `null-coalescing operator` for this case, and if `view count` will be equals to `null`, it will be set to `0`.